requirements.txt
numpy
cv2
matplotlib
absl-py==0.11.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613038019788/work
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
cachetools==4.2.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
chardet==4.0.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1613253194046/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dicom-contour==2
entrypoints==0.3
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.10.0
idna==2.10
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877486026/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1617121109687/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213516947/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
Keras-Preprocessing==1.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1603996595157/work
Markdown==3.3.3
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-base_1603356257853/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601914921407/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
notebook @ file:///C:/ci/notebook_1616443616158/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1611178799982/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102427207/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1609786872067/work
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1618088486455/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydicom==2.1.2
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600123688363/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets @ file:///C:/ci/pywavelets_1601658407053/work
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1616775094278/work
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.16.2
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1612469765256/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///C:/ci/six_1605205426665/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torch==1.7.1
torchvision==0.8.2
tornado @ file:///C:/ci/tornado_1606935947090/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xlrd @ file:///tmp/build/80754af9/xlrd_1608072521494/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
